+++ packages/react-native-sdk/src/components/call/CallControls/CallControlsButton.tsx
@@ -19,6 +19,10 @@ interface CallControlsButtonProps {
    */
   color?: string;
   /**
+   * Boolean to enable/disable the button
+   */
+  disabled?: boolean;
+  /**
    * Style to the Pressable button.
    */
   style?: StyleProp<ViewStyle>;
@@ -30,6 +34,11 @@ interface CallControlsButtonProps {
    * Accessibility label for the button.
    */
   testID?: string;
+  /**
+   * Invoked on mount and layout changes with
+   * {nativeEvent: { layout: {x, y, width, height}}}.
+   */
+  onLayout?: View['props']['onLayout'];
 }
 
 const DEFAULT_ICON_SIZE = theme.icon.md;
@@ -38,7 +47,16 @@ const DEFAULT_BUTTON_SIZE = theme.button.sm;
 export const CallControlsButton = (
   props: React.PropsWithChildren<CallControlsButtonProps>,
 ) => {
-  const { onPress, children, color, style, svgContainerStyle, testID } = props;
+  const {
+    onPress,
+    children,
+    disabled,
+    color,
+    style,
+    svgContainerStyle,
+    testID,
+    onLayout,
+  } = props;
 
   const pressableStyle: PressableProps['style'] = ({ pressed }) => [
     DEFAULT_BUTTON_SIZE,
@@ -47,18 +65,19 @@ export const CallControlsButton = (
       backgroundColor: color,
       opacity: pressed ? 0.2 : 1,
     },
-    style ? style : null,
+    style,
+    disabled ? styles.disabledStyle : null,
   ];
 
   return (
-    <Pressable style={pressableStyle} onPress={onPress} testID={testID}>
-      <View
-        style={[
-          styles.svgContainerStyle,
-          DEFAULT_ICON_SIZE,
-          svgContainerStyle ?? null,
-        ]}
-      >
+    <Pressable
+      disabled={disabled}
+      style={pressableStyle}
+      onPress={onPress}
+      testID={testID}
+      onLayout={onLayout}
+    >
+      <View style={[DEFAULT_ICON_SIZE, svgContainerStyle ?? null]}>
         {children}
       </View>
     </Pressable>
@@ -71,6 +90,19 @@ const styles = StyleSheet.create({
     borderWidth: 1,
     borderColor: theme.light.content_bg,
     alignItems: 'center',
+    // For iOS
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 8,
+
+    // For android
+    elevation: 6,
   },
-  svgContainerStyle: {},
+  disabledStyle: {
+    backgroundColor: theme.light.disabled,
+  },
 });