+++ packages/react-native-sdk/src/components/call/CallControls/HangupCallButton.tsx
@@ -1,10 +1,10 @@
 import React, { useCallback } from 'react';
 import { CallControlsButton } from './CallControlsButton';
 import { theme } from '../../../theme';
-import { StyleProp, StyleSheet, ViewStyle } from 'react-native';
+import { StyleProp, ViewStyle } from 'react-native';
 import { PhoneDown } from '../../../icons';
 import { ButtonTestIds } from '../../../constants/TestIds';
-import { useCall, useCallCallingState } from '@stream-io/video-react-bindings';
+import { useCall, useCallStateHooks } from '@stream-io/video-react-bindings';
 import { CallingState } from '@stream-io/video-client';
 
 /**
@@ -16,18 +16,30 @@ export type HangUpCallButtonProps = {
    * @returns void
    */
   onPressHandler?: () => void;
-
+  /**
+   * Handler to be called when the call is hanged up.
+   *
+   * Note: If the `onPressHandler` is passed this handler will not be executed.
+   */
+  onHangUpCall?: () => void;
   /**
    * Style of the Button Container
    */
   style?: StyleProp<ViewStyle>;
 };
 
+/**
+ * Button to hangup a call.
+ *
+ * Mostly calls call.leave() internally.
+ */
 export const HangUpCallButton = ({
   onPressHandler,
+  onHangUpCall,
   style,
 }: HangUpCallButtonProps) => {
   const call = useCall();
+  const { useCallCallingState } = useCallStateHooks();
   const callingState = useCallCallingState();
 
   const hangUpCallHandler = useCallback(async () => {
@@ -40,6 +52,9 @@ export const HangUpCallButton = ({
         return;
       }
       await call?.leave();
+      if (onHangUpCall) {
+        onHangUpCall();
+      }
     } catch (error) {
       console.error('Error leaving call:', error);
     }
@@ -50,25 +65,10 @@ export const HangUpCallButton = ({
     <CallControlsButton
       onPress={hangUpCallHandler}
       color={theme.light.error}
-      style={[styles.button, { shadowColor: theme.light.error }, style]}
+      style={[{ shadowColor: theme.light.error }, style]}
       testID={ButtonTestIds.HANG_UP_CALL}
     >
       <PhoneDown color={theme.light.static_white} />
     </CallControlsButton>
   );
 };
-
-const styles = StyleSheet.create({
-  button: {
-    // For iOS
-    shadowOffset: {
-      width: 0,
-      height: 6,
-    },
-    shadowOpacity: 0.37,
-    shadowRadius: 7.49,
-
-    // For android
-    elevation: 6,
-  },
-});