+++ packages/react-native-sdk/src/components/participants/LocalParticipantView/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { StyleProp, StyleSheet, View, ViewStyle } from 'react-native';
-import { useLocalParticipant } from '@stream-io/video-react-bindings';
+import { useCallStateHooks } from '@stream-io/video-react-bindings';
 import { SfuModels } from '@stream-io/video-client';
-import { Z_INDEX, LOCAL_VIDEO_VIEW_STYLE } from '../../../constants';
+import { LOCAL_VIDEO_VIEW_STYLE, Z_INDEX } from '../../../constants';
 import { ComponentTestIds } from '../../../constants/TestIds';
 import { VideoSlash } from '../../../icons';
 import { useMediaStreamManagement } from '../../../providers';
@@ -45,6 +45,7 @@ export interface LocalParticipantViewProps {
  */
 export const LocalParticipantView = (props: LocalParticipantViewProps) => {
   const { layout = 'floating' } = props;
+  const { useLocalParticipant } = useCallStateHooks();
   const localParticipant = useLocalParticipant();
   const { isCameraOnFrontFacingMode } = useMediaStreamManagement();
   // it takes a few milliseconds for the camera stream to actually switch