+++ sample-apps/react-native/dogfood/src/components/ActiveCall.tsx
@@ -1,14 +1,18 @@
 import React, { useEffect, useRef, useState } from 'react';
 import {
-  CallContentProps,
-  CallControls,
-  CallControlsType,
   CallingState,
-  ParticipantsInfoBadge,
   CallContent,
   useCall,
   useIncallManager,
   theme,
+  ReactionButton,
+  ChatButton,
+  ToggleVideoPublishingButton,
+  ToggleAudioPublishingButton,
+  ToggleCameraFaceButton,
+  HangUpCallButton,
+  CallTopView,
+  ChatButtonProps,
 } from '@stream-io/video-react-native-sdk';
 import { ActivityIndicator, StyleSheet, View } from 'react-native';
 import { appTheme } from '../theme';
@@ -16,21 +20,29 @@ import {
   SafeAreaView,
   useSafeAreaInsets,
 } from 'react-native-safe-area-context';
-import { ActiveCallNotification } from './ActiveCallNotification';
-import { ParticipantsLayoutSwitchButton } from './ParticipantsLayoutButton';
+import { ParticipantsInfoList } from './ParticipantsInfoList';
+import { Z_INDEX } from '../constants';
 
-type ActiveCallProps = CallControlsType;
-
-type Layout = CallContentProps['mode'];
+type ActiveCallProps = {
+  chatButton?: ChatButtonProps;
+  onHangupCallHandler?: () => void;
+  onBackPressed?: () => void;
+};
 
 export const ActiveCall = ({
   chatButton,
-  hangUpCallButton,
+  onBackPressed,
+  onHangupCallHandler,
 }: ActiveCallProps) => {
   const call = useCall();
   const activeCallRef = useRef(call);
   activeCallRef.current = call;
-  const [selectedLayout, setSelectedLayout] = useState<Layout>('grid');
+  const [isCallParticipantsVisible, setIsCallParticipantsVisible] =
+    useState<boolean>(false);
+
+  const onOpenCallParticipantsInfo = () => {
+    setIsCallParticipantsVisible(true);
+  };
 
   useEffect(() => {
     return () => {
@@ -45,7 +57,7 @@ export const ActiveCall = ({
    */
   useIncallManager({ media: 'video', auto: true });
 
-  const { bottom, top } = useSafeAreaInsets();
+  const { bottom } = useSafeAreaInsets();
 
   if (!call) {
     return <ActivityIndicator size={'large'} style={StyleSheet.absoluteFill} />;
@@ -53,22 +65,35 @@ export const ActiveCall = ({
 
   return (
     <SafeAreaView style={styles.container} edges={['top', 'left', 'right']}>
-      <ActiveCallNotification />
-      <View style={[styles.icons, { top }]}>
-        <ParticipantsLayoutSwitchButton
-          selectedLayout={selectedLayout}
-          setSelectedLayout={setSelectedLayout}
+      <View style={styles.container}>
+        <CallTopView
+          onBackPressed={onBackPressed}
+          onParticipantInfoPress={onOpenCallParticipantsInfo}
         />
-        <ParticipantsInfoBadge />
+        <CallContent />
       </View>
-      <CallContent mode={selectedLayout} />
-      <CallControls
-        chatButton={chatButton}
-        hangUpCallButton={hangUpCallButton}
+      {/* Since we want the chat and the reaction button the entire call controls is customized */}
+      <View
         style={[
           styles.callControlsWrapper,
-          { paddingBottom: Math.max(bottom, appTheme.spacing.lg) },
+          {
+            paddingBottom: Math.max(bottom, appTheme.spacing.lg),
+          },
         ]}
+      >
+        <ReactionButton />
+        <ChatButton
+          onPressHandler={chatButton?.onPressHandler}
+          unreadBadgeCount={chatButton?.unreadBadgeCount}
+        />
+        <ToggleVideoPublishingButton />
+        <ToggleAudioPublishingButton />
+        <ToggleCameraFaceButton />
+        <HangUpCallButton onPressHandler={onHangupCallHandler} />
+      </View>
+      <ParticipantsInfoList
+        isCallParticipantsInfoVisible={isCallParticipantsVisible}
+        setIsCallParticipantsInfoVisible={setIsCallParticipantsVisible}
       />
     </SafeAreaView>
   );
@@ -77,7 +102,6 @@ export const ActiveCall = ({
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: appTheme.colors.static_grey,
   },
   icons: {
     position: 'absolute',
@@ -88,7 +112,10 @@ const styles = StyleSheet.create({
     zIndex: appTheme.zIndex.IN_FRONT,
   },
   callControlsWrapper: {
-    paddingTop: appTheme.spacing.lg,
-    paddingHorizontal: appTheme.spacing.sm,
+    flexDirection: 'row',
+    justifyContent: 'space-evenly',
+    paddingVertical: theme.padding.md,
+    zIndex: Z_INDEX.IN_FRONT,
+    backgroundColor: appTheme.colors.static_grey,
   },
 });