+++ sample-apps/react/react-dogfood/components/CallLayout/CallParticipantsView.tsx
@@ -1,11 +1,11 @@
 import {
-  Call,
   DefaultParticipantViewUI,
   ParticipantView,
-  useParticipants,
+  useCallStateHooks,
 } from '@stream-io/video-react-sdk';
 
-export const CallParticipantsView = (props: { call: Call }) => {
+export const CallParticipantsView = () => {
+  const { useParticipants } = useCallStateHooks();
   const participants = useParticipants();
   const grid = `str-video__grid-${participants.length || 1}`;
   return (